Update README.md
Brushed up the README file a bit. Still things to add, though. Should provide all relevant information to someone looking at the project (this is the first page they see). Where things get very detailed (e.g. with the description of the constraint formats), it is probably good to create a separate document and link to there.